chore: hard delete sync jobs when a sync is deleted
Sync jobs table is growing rapidely and could become a problem at some
point. The only advantage of doing a soft delete is debugging but we
rarely need it. We could also have a more complex 2 steps delete
(immediate soft + hard later) but I don't think it is worth it.